split expansion functions into separate files
