+++ Makefile
@@ -6,7 +6,7 @@
 #    By: jhor <jhor@student.42kl.edu.my>            +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/09/03 18:00:38 by jhor              #+#    #+#              #
-#    Updated: 2025/11/28 14:54:57 by jhor             ###   ########.fr        #
+#    Updated: 2025/12/01 11:24:27 by jhor             ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -26,8 +26,14 @@ src/ast/helper/error_syntax.c src/ast/tokenization/helper_f.c \
 src/ast/parse_pipe.c ast_visualizer/AST.c src/ast/helper_ast.c \
 src/ast/helper/utils.c  src/ast/helper/utils2.c src/ast/parse_word.c \
 src/ast/parse_redir.c src/ast/parse_argument.c src/ast/parse_component.c \
-src/ast/parse_simple_command.c src/ast/parsing.c src/heredoc/heredoc_active.c \
-src/set_envp.c src/expansion/core.c src/expansion/instructor.c \
+src/ast/parse_simple_command.c \
+src/ast/parsing.c src/heredoc/heredoc_active.c \
+src/set_envp.c src/expansion/expand_start.c src/expansion/instructor.c \
+src/expansion/remove_quotes.c src/expansion/cmd_argv.c src/expansion/copy_split.c \
+src/expansion/expand_check.c src/expansion/expand_split.c \
+src/expansion/expand_split_helper.c src/expansion/assign_mark.c \
+src/expansion/token_double.c src/expansion/token_single.c \
+src/expansion/extract_expand.c \
 
 OBJ = $(SRC:.c=.o)
